bot/requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
python-dotenv==0.21.0
emoji==2.2.0
